+++ src/main/java/com/example/demo/Board.java
@@ -1,4 +1,4 @@
-package main;
+package com.example.demo;
 
 import java.util.List;
 import java.util.Map;
