+++ src/main/java/com/example/demo/CornerNode.java
@@ -1,4 +1,4 @@
-package main;
+package com.example.demo;
 
 public class CornerNode extends Node {
 }
