+++ src/main/java/com/example/demo/NormalNode.java
@@ -1,4 +1,4 @@
-package main;
+package com.example.demo;
 
 public class NormalNode extends Node {
