+++ src/main/java/com/example/demo/Player.java
@@ -1,4 +1,4 @@
-package main;
+package com.example.demo;
 
 import java.util.ArrayList;
 import java.util.List;
