+++ src/main/java/com/example/demo/RandomYut.java
@@ -1,4 +1,4 @@
-package main;
+package com.example.demo;
 
 import java.util.Random;
