+++ src/main/java/com/example/demo/TetragonalBoard.java
@@ -1,13 +1,13 @@
-package main;
-//5각형
+package com.example.demo;
+//4각형
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-public class PentagonalBoard extends Board {
+public class TetragonalBoard extends Board {
 
-    public PentagonalBoard() {
+    public TetragonalBoard() {
         nodes = new ArrayList<>();
         edges = new HashMap<>();
     }
@@ -15,30 +15,29 @@ public class PentagonalBoard extends Board {
     @Override
     public void createNodes() {
 
-        for (int i = 0; i < 37; i++)
+        for (int i = 0; i < 30; i++)
             nodes.add(new Node(i));
     }
 
     @Override
     public void createEdges() {
 
-        for (int i = 0; i < 36; i++) {
+        for (int i = 0; i < 29; i++) // 지름길 아닐때는 그냥 +1
+        {
             edges.put(i, List.of(i + 1));
         }
 
-        edges.put(27, List.of(25));
-        edges.put(29, List.of(25));
+        // 지름길 -> 위의 결과보다 더 빠른 순서로 저장
 
-        edges.put(31, List.of(30));
-        edges.put(30, List.of(25));
+        edges.put(22, List.of(21));
+        edges.put(21, List.of(20));
 
-        // 중심점 25에서 멈추지 않았을때
-        edges.put(25, List.of(32));
-        edges.put(33, List.of(20));
+        edges.put(24, List.of(23));
+        edges.put(23, List.of(20));
 
-        // 중심점 25에서 멈췄을때
-        edges.put(34, List.of(35));
-        edges.put(35, List.of(36));
+        // 딱 중심점에서 멈추지 않았을때
+        edges.put(26, List.of(15));
+        edges.put(20, List.of(25));
 
     }
 
@@ -50,45 +49,45 @@ public class PentagonalBoard extends Board {
     @Override
     public void movePosition(Piece myPiece, Integer yutValue) {
 
-        if (myPiece.isFinished())
+        if (myPiece.isFinished()) // 말이 끝까지 갔으면 더 움직일 수 없음
             return;
 
-        int position = myPiece.getPosition();
-        Node currentNode = nodes.get(position);
-        currentNode.remove(myPiece);
+        int position = myPiece.getPosition(); // 말의 현재 위치 가져옴
+        Node currentNode = nodes.get(position); // 말이 현재 위치해 있는 노드 가져옴
+        currentNode.remove(myPiece); // 현재 위치해 있는 노드에서 말 삭제
 
+        // 빽도
         if (yutValue == -1) {
-            int prev = myPiece.popPreviousPosition();
-            if (prev != -1) {
+            int prev = myPiece.popPreviousPosition(); // 말이 지나온 경로 중 가장 최근 위치
+            if (prev != -1) // 뒤로 갈 수 있을 때
+            {
                 position = prev;
-                System.out.println("빽도");
-            } else {
-                System.out.println("뒤로 갈 수 없음");
+                System.out.println("빽도"); // 테스트용으로 써본겁니다
+            } else // 시작지점일때
+            {
+                System.out.println("뒤로 갈 수 없음"); // 테스트용으로 써본겁니다
             }
             myPiece.setPosition(position);
             nodes.get(position).add(myPiece);
             return;
         }
 
-        if (position == 25) // 시작 위치가 25일때(25에서 딱 멈췄을때)
+        if (position == 5) // 시작 위치가 5일때(5에서 딱 멈췄을때)
         {
             myPiece.pushPreviousPosition(position);
-            position = 34;
+            position = 22;
             yutValue--;
-        } else if (position == 5) // 시작 위치가 25일때(25에서 딱 멈췄을때)
+        }
+        if (position == 10) // 시작 위치가 10일때(10에서 딱 멈췄을때)
         {
             myPiece.pushPreviousPosition(position);
-            position = 26;
+            position = 24;
             yutValue--;
-        } else if (position == 10) // 시작 위치가 25일때(25에서 딱 멈췄을때)
+        }
+        if (position == 20) // 시작 위치가 20일때(20에서 딱 멈췄을때)
         {
             myPiece.pushPreviousPosition(position);
-            position = 28;
-            yutValue--;
-        } else if (position == 15) // 시작 위치가 25일때(25에서 딱 멈췄을때)
-        {
-            myPiece.pushPreviousPosition(position);
-            position = 31;
+            position = 27;
             yutValue--;
         }
 
@@ -99,31 +98,32 @@ public class PentagonalBoard extends Board {
                 // 종점(시작점)을 통과하거나 이동할 곳이 없으면 승리 처리
                 System.out.println("승리");
                 myPiece.finish();
-                position = 36; // 명시적으로 승리 위치 지정
+                position = 29; // 명시적으로 승리 위치 지정
                 break;
             }
 
             myPiece.pushPreviousPosition(position);
             position = nextPosition.get(0);
 
-            if (position == 36) {
+            if (position == 29) {
                 System.out.println("승리");
                 myPiece.finish();
                 break;
             }
         }
 
-        // 잡기
-        Node nextNode = nodes.get(position);
-        List<Piece> pieces = new ArrayList<>(nextNode.getOwnedPieces());
+        // 잡기 & 그룹핑
+        Node nextNode = nodes.get(position); // 말이 도착할 위치에 다른 말이 있는지 알기 위해 ..
+        List<Piece> pieces = new ArrayList<>(nextNode.getOwnedPieces()); // 말이 도착할 위치에 있는 모든 말들의 리스트
 
         for (int i = 0; i < pieces.size(); i++) {
             Piece opponentPiece = pieces.get(i);
-            if (opponentPiece.getOwnerId() != myPiece.getOwnerId()) {
+            if (opponentPiece.getOwnerId() != myPiece.getOwnerId()) // 같은 플레이어의 말이 아닐때 -> 잡기
+            {
                 nextNode.remove(pieces.get(i));
                 opponentPiece.setPosition(0);
                 nodes.get(0).add(opponentPiece);
-                System.out.println("상대 팀 말 잡음!");
+                System.out.println("상대 팀 말 잡음! "); // 테스트용으로 써본겁니다
             } else if (opponentPiece.getOwnerId() == myPiece.getOwnerId()) // 같은 플레이어 말일때 -> 그룹핑
             {
                 myPiece.grouping(opponentPiece);
@@ -145,5 +145,4 @@ public class PentagonalBoard extends Board {
         }
 
     }
-
 }
