+++ src/main/java/com/example/demo/Yut.java
@@ -1,4 +1,4 @@
-package main;
+package com.example.demo;
 
 abstract class Yut {
     public abstract Integer getResult();
