--- src/test/java/GameTest/ServiceTest.java
@@ -1,9 +1,0 @@
-package GameTest;
-
-import static org.junit.jupiter.api.Assertions.*;
-class ServiceTest {
-
-    @org.junit.jupiter.api.Test
-    void startGame() {
-    }
-}
