+++ src/test/java/com/example/demo/controller.java
@@ -1,0 +1,4 @@
+package com.example.demo;
+
+public class controller {
+}
