+++ src/test/java/com/example/demo/repository.java
@@ -1,0 +1,4 @@
+package com.example.demo;
+
+public class repository {
+}
