+++ src/test/java/com/example/demo/service.java
@@ -1,0 +1,4 @@
+package com.example.demo;
+
+public class service {
+}
